Experience is now handled internally as a double.

+++ src/com/herocraftonline/dev/heroes/HEntityListener.java
@@ -41,12 +41,12 @@ public class HEntityListener extends EntityListener {
             // 5% of the next level's experience requirement
             // Experience loss can't reduce level
             Hero heroDefender = plugin.getHeroManager().getHero((Player) defender);
-            int exp = heroDefender.getExperience();
+            double exp = heroDefender.getExperience();
             int level = prop.getLevel(exp);
             if (level < prop.maxLevel) {
-                int currentLevelExp = prop.getExperience(level);
-                int nextLevelExp = prop.getExperience(level + 1);
-                int expLoss = (int) ((nextLevelExp - currentLevelExp) * prop.expLoss);
+                int currentLevelExp = (int)prop.getExperience(level);
+                int nextLevelExp = (int)prop.getExperience(level + 1);
+                double expLoss = (nextLevelExp - currentLevelExp) * prop.expLoss;
                 if (exp - expLoss < currentLevelExp) {
                     expLoss = exp - currentLevelExp;
                 }
